Update waistAngles function to return the second result in the "opposite direction".
This isn't guaranteed to avoid joint angle limits but it should be better more times than not. We will need to do more thorough checking to make sure that a suggested angle that is outside of one limit isn't inside of the other limit (e.g. 330 degrees outside limits but -30 degrees is not)