:boom: :tada: Major rewrite of the inverse kinematics.
All robot related parameters are pulled up to the top level function call. solveElbow only returns a single value now. The caller is responsible for dealing with the opposite configuration (-elbow). solveElbow no longer needs to do any checking on the cosTheta value and instead returns a nan for target positions out of reach of the manipulator.